Migrate build tooling from @crxjs/vite-plugin to WXT
- Replace vite.config.ts + manifest.config.ts with wxt.config.ts (srcDir: 'src')
- Move entrypoints under src/entrypoints/ per WXT convention
- Switch to defineBackground (auto-imported); options entry uses meta tag
  manifest.open_in_tab=true to keep full-page behavior previously provided by
  options_page
- Relocate icons to public/icon/{16,48,128}.png for WXT auto-discovery
- Drop @crxjs/vite-plugin, vite, @vitejs/plugin-vue, vue-tsc, vitest, and
  Histoire (incl. story files); pin wxt 0.20.25 + @wxt-dev/module-vue 1.0.3
- tsconfig.json now extends ./.wxt/tsconfig.json; add shims-vue.d.ts so
  src/App.vue can be imported from the entrypoint
- Replace dev/prod toggle: manifest function uses
  command === 'serve' || mode === 'development' to flip name to [DEV] OmniATP
  and omnibox keyword to atd
