--- histoire.config.ts
@@ -1,7 +1,0 @@
-import { defineConfig } from 'histoire'
-import { HstVue } from '@histoire/plugin-vue'
-
-export default defineConfig({
-  plugins: [HstVue()],
-  setupFile: 'histoire-setup.ts',
-})
