+++ package.json
@@ -4,10 +4,13 @@
   "version": "0.0.3",
   "type": "module",
   "scripts": {
-    "dev": "vite",
-    "build": "vue-tsc && vite build",
-    "preview": "vite preview",
-    "story:dev": "histoire dev",
+    "dev": "wxt",
+    "dev:firefox": "wxt -b firefox",
+    "build": "wxt build",
+    "build:firefox": "wxt build -b firefox",
+    "zip": "wxt zip",
+    "compile": "wxt prepare && tsc --noEmit",
+    "postinstall": "wxt prepare",
     "lint:fix": "eslint --fix . && prettier -w ."
   },
   "dependencies": {
@@ -21,27 +24,22 @@
     "vue": "3.5.25"
   },
   "devDependencies": {
-    "@crxjs/vite-plugin": "2.3.0",
     "@eslint/js": "9.39.2",
-    "@histoire/plugin-vue": "0.17.17",
     "@tailwindcss/forms": "0.5.10",
     "@tailwindcss/vite": "4.1.18",
     "@types/chrome": "0.1.32",
-    "@vitejs/plugin-vue": "6.0.6",
+    "@wxt-dev/module-vue": "1.0.3",
     "eslint": "9.39.2",
     "eslint-config-prettier": "10.1.8",
     "eslint-plugin-vue": "10.6.2",
     "globals": "16.5.0",
-    "histoire": "0.17.17",
     "postcss": "8.5.6",
     "prettier": "3.7.4",
     "tailwindcss": "4.1.18",
     "typescript": "5.9.3",
     "typescript-eslint": "8.50.0",
-    "vite": "7.3.2",
-    "vitest": "4.0.15",
     "vue-eslint-parser": "10.2.0",
-    "vue-tsc": "3.1.8"
+    "wxt": "0.20.25"
   },
   "packageManager": "pnpm@10.26.0"
 }
