--- src/components/LoadingSpinner.story.vue
@@ -1,11 +1,0 @@
-<script setup lang="ts">
-import LoadingSpinner from './LoadingSpinner.vue'
-</script>
-
-<template>
-  <Story title="components/LoadingSpinner">
-    <div class="p-4">
-      <LoadingSpinner />
-    </div>
-  </Story>
-</template>
