+++ src/entrypoints/options/index.html
@@ -4,10 +4,11 @@
     <meta charset="UTF-8" />
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <meta name="manifest.open_in_tab" content="true" />
     <title>Options - OmniATP</title>
   </head>
   <body>
     <div id="app"></div>
-    <script type="module" src="options.ts"></script>
+    <script type="module" src="./main.ts"></script>
   </body>
 </html>
