--- src/vite-env.d.ts
@@ -1,2 +1,0 @@
-/// <reference types="vite/client" />
-/// <reference types="@histoire/plugin-vue/components" />
