+++ tsconfig.node.json
@@ -6,5 +6,5 @@
     "allowSyntheticDefaultImports": true,
     "resolveJsonModule": true
   },
-  "include": ["vite.config.ts", "manifest.config.ts", "package.json"]
+  "include": ["wxt.config.ts", "package.json"]
 }
